Fix #20
In a recent dnscrypt-proxy update the privileges are also dropped when
using the -check command. Because the SRM web UI and the cgi script is
running as the 'system' user, when running dnscrypt-proxy we do not
have the privileges to the system calls to drop privileges further.

The fix here is just to disable the syntax check of the config files

Not that it matters now but the netprobe-timeout=0 is no longer needed:
> Netprobes and listening sockets are now ignored when the -list, -list-all, -show-certs or -check command-line switches are used

https://github.com/publicarray/spksrc/issues/20